third_party/WebKit/LayoutTests/fast/dom/Document/scrollingElement-noStyleUpdate-strict.html

<!DOCTYPE html>
<style>
</style>
<html>
    <body>
    </body>
</html>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/run-after-display.js"></script>
<script>
    description("Verify that document.scrollingElement doesn't force a style update or layout in strict mode");
    setPrintTestResultsLazily();

    document.body.offsetTop;
    shouldBe("internals.needsLayoutCount()", "0");

    document.body.style.padding = "10px";
    shouldBe("document.scrollingElement", "document.documentElement");
    debug("Verifying style update hadn't been triggered");
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");
    debug("Verifying layout hasn't been triggered");
    shouldBe("internals.needsLayoutCount()", "3");
</script>
