third_party/WebKit/LayoutTests/fast/dom/Document/scrollingElement-noStyleUpdate-strict-expected.txt

Verify that document.scrollingElement doesn't force a style update or layout in strict mode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.needsLayoutCount() is 0
PASS document.scrollingElement is document.documentElement
Verifying style update hadn't been triggered
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
Verifying layout hasn't been triggered
PASS internals.needsLayoutCount() is 3
PASS successfullyParsed is true

TEST COMPLETE
